// src/Service/mipg/dto/get-mipg.dto.ts
export class GetMipgDto {
  user?: string;
  ref?: string;
  terminalid?: number;
  status?: number;
  readonly amount?: number;
  readonly payload?: string;
  readonly callbackurl?: string;
  invoiceid?: string;
  userinvoice?: string;
  readonly cardno?: string;
  readonly password?: string;
  readonly nationalcode?: string;
  readonly cardpassword?: string;
  devicetype?: string;
  paytype?: number;
  karmozd?: number;
  wagetype?: number;
  isdirect?: boolean;
}